```text feat: Add Mitosis library for building React and Vue components
This commit adds the Mitosis library to the project, which allows for building components that can be used in both React and Vue. Mitosis is a frontend compiler that can compile components into code for multiple frameworks, including React and Vue. It provides a set of APIs and supports features like state management, styling, and flow control. The library is integrated into the project and configured to support both React and Vue targets. This will enable the development of components that can be used in both frameworks, reducing the need for separate adaptation code.